chore(uptime): Move EAP testcase to testutils
We want to use this in `getsentry`, so moving this
